Add a simple coin-toss probabilistic model example.
The example illustrates the way the model is set up.
It shows how observed values (data) is passed into NumPyro samplers
and how determined parameters are extracted.

Additionally two sections illustrate the way to set up a 'potential' function
using log(p(x|data) and how to compute the gradients of this preliminary
stge of a potential.