Constraining parameters and transforming gradients
Added examples on how to constrain positions (parameter vectors)
from |R^n onto the support of the model distribution and how
to determine the Jacobi matrix of such a transform.

Added also a part on how to obtain the gradient on constrained
space and to transform it back to unconstrained space.